requirements.txt
# MLOps ClearML Project Dependencies
# Core ML
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.5.0
scikit-learn>=1.2.0

# ClearML
clearml>=1.14.0

# Gradio for UI
gradio>=4.0.0

# Dataset download
kagglehub>=0.2.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
